Move ISIS-related live data classes to ISIS subdirectory
