+++ Framework/LiveData/CMakeLists.txt
@@ -2,16 +2,14 @@ set ( SRC_FILES
 	src/ADARA/ADARAPackets.cpp
 	src/ADARA/ADARAParser.cpp
 	src/FakeEventDataListener.cpp
-	src/FakeISISHistoDAE.cpp
 	src/FileEventDataListener.cpp
+	src/ISIS/DAE/idc.cpp
+	src/ISIS/DAE/isisds_command.cpp
 	src/ISIS/FakeISISEventDAE.cpp
-	src/ISISHistoDataListener.cpp
-	src/ISISLiveEventDataListener.cpp
+	src/ISIS/FakeISISHistoDAE.cpp
+	src/ISIS/ISISHistoDataListener.cpp
+	src/ISIS/ISISLiveEventDataListener.cpp
 	src/LiveDataAlgorithm.cpp
-	src/LoadDAE/idc.cpp
-	src/LoadDAE/idc.h
-	src/LoadDAE/isisds_command.cpp
-	src/LoadDAE/isisds_command.h
 	src/LoadLiveData.cpp
 	src/MonitorLiveData.cpp
 	src/SNSLiveEventDataListener.cpp
@@ -19,8 +17,8 @@ set ( SRC_FILES
 	src/TOPAZLiveEventDataListener.cpp
 )
 
-set ( SRC_UNITY_IGNORE_FILES src/LoadDAE/idc.cpp
-        src/LoadDAE/isisds_command.cpp)
+set ( SRC_UNITY_IGNORE_FILES src/ISIS/DAE/idc.cpp
+        src/ISIS/DAE/isisds_command.cpp)
 
 set ( INC_FILES
 	inc/MantidLiveData/ADARA/ADARA.h
@@ -28,18 +26,20 @@ set ( INC_FILES
 	inc/MantidLiveData/ADARA/ADARAParser.h
 	inc/MantidLiveData/Exception.h
 	inc/MantidLiveData/FakeEventDataListener.h
-	inc/MantidLiveData/FakeISISHistoDAE.h
 	inc/MantidLiveData/FileEventDataListener.h
 	inc/MantidLiveData/ISIS/FakeISISEventDAE.h
+	inc/MantidLiveData/ISIS/FakeISISHistoDAE.h
 	inc/MantidLiveData/ISIS/TCPEventStreamDefs.h
-	inc/MantidLiveData/ISISHistoDataListener.h
-	inc/MantidLiveData/ISISLiveEventDataListener.h
+	inc/MantidLiveData/ISIS/ISISHistoDataListener.h
+	inc/MantidLiveData/ISIS/ISISLiveEventDataListener.h
 	inc/MantidLiveData/LiveDataAlgorithm.h
 	inc/MantidLiveData/LoadLiveData.h
 	inc/MantidLiveData/MonitorLiveData.h
 	inc/MantidLiveData/SNSLiveEventDataListener.h
 	inc/MantidLiveData/StartLiveData.h
 	inc/MantidLiveData/TOPAZLiveEventDataListener.h
+	src/ISIS/DAE/idc.h
+	src/ISIS/DAE/isisds_command.h
 )
 
 set ( TEST_FILES
