+++ Framework/LiveData/inc/MantidLiveData/ISIS/ISISHistoDataListener.h
